kodo: add test case for reuse of incomplete decoders
Since some usecases might return decoders to the factory, even if
decoding is incomplete (e.g. after a timeout), it should be tested if
the reuse of such decoders is correct.

This patch adds a test case that 1) build coders from a factory; 2)
encode and decode packets with a loss probability; 3) discontinues
an incomplete decoding with a certain probability.

Signed-off-to: Steinwurf ApS
Signed-off-by: Martin Hundebøll <martin@hundeboll.net>