kodo: support partial decoding
To allow decoders to extract decoded symbols before all symbols has been
received, the decoding layers needs to keep track of the maximum
coefficient index.

This patch adds a member variable to hold the maximum index, which is
updated whenever a new symbol with more none-zero coefficients is added.
To allow users to detect when decoded symbols are available, a public
function is added to test wether the block has rank equal to the maximum
coefficient index.

Signed-off-by: Martin Hundebøll <martin@hundeboll.net>
Signed-off-to: Steinwurf ApS <contact@steinwurf.com>